+++ app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @open-tomato/app
 
+## 0.0.3
+
+### Patch Changes
+
+- @open-tomato/ui-components@0.8.2
+
 ## 0.0.2
 
 ### Patch Changes